Move country loading code into subpackage
