+++ examples/contacts-example/src/main/java/de/saxsys/mvvmfx/examples/contacts/model/countries/ISO3166_2_EntryEntity.java
@@ -1,4 +1,4 @@
-package de.saxsys.mvvmfx.examples.contacts.model;
+package de.saxsys.mvvmfx.examples.contacts.model.countries;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
